requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
waitress==2.1.2
requests==2.31.0
beautifulsoup4==4.12.2
python-dotenv==1.0.0
pydantic==2.5.2
aiohttp==3.9.1
